+++ src/fidesops/ops/models/privacy_request.py
@@ -6,7 +6,7 @@ import json
 import logging
 from datetime import datetime, timedelta
 from enum import Enum as EnumType
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Union
 
 from celery.result import AsyncResult
 from fideslib.cryptography.cryptographic_util import hash_with_salt
@@ -212,6 +212,14 @@ class PrivacyRequest(Base):  # pylint: disable=R0904
     paused_at = Column(DateTime(timezone=True), nullable=True)
     identity_verified_at = Column(DateTime(timezone=True), nullable=True)
     due_date = Column(DateTime(timezone=True), nullable=True)
+
+    @property
+    def days_left(self: PrivacyRequest) -> Union[int, None]:
+        if self.due_date is None:
+            return None
+
+        delta = self.due_date.date() - datetime.utcnow().date()
+        return delta.days
 
     @classmethod
     def create(cls, db: Session, *, data: Dict[str, Any]) -> FidesBase: