+++ src/fidesops/ops/schemas/privacy_request.py
@@ -134,7 +134,6 @@ class RowCountRequest(BaseSchema):
 
 
 class CheckpointActionRequiredDetails(CheckpointActionRequired):
-
     collection: Optional[str] = None  # type: ignore
 
 
@@ -166,6 +165,7 @@ class PrivacyRequestResponse(BaseSchema):
     policy: PolicySchema
     action_required_details: Optional[CheckpointActionRequiredDetails] = None
     resume_endpoint: Optional[str]
+    days_left: Optional[int]
 
     class Config:
         """Set orm_mode and use_enum_values"""